Follow walls
Right now the camera confuses things a little - since the way the
camera is set up means it ALWAYS sees red, green or blue - it’s never
neutral. But, it should follow a corridor, and if one wall disappears
then it hugs the remaining one. If it’s meant to be turning next then
it’ll hug the one that disappeared instead